+++ app/pb_kits/playbook/pb_installer/docs/_installer_default.jsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import Installer from '../_installer.jsx'
+
+const InstallerDefault = () => (
+  <div>
+    <Installer projectName="JD Installations LLC" />
+    <br />
+    <br />
+    <Installer
+        link="#"
+        projectName="JD Installations LLC"
+    />
+  </div>
+)
+
+export default InstallerDefault