+++ app/pb_kits/playbook/pb_list/docs/_list_dark.jsx
@@ -6,12 +6,7 @@ const ListDark = () => {
   return (
     <>
       <List
-          borderless
           dark
-          layout="right"
-          ordered
-          size="large"
-          xpadding
       >
         <ListItem>{'Item 1'}</ListItem>
         <ListItem>{'Item 2'}</ListItem>