+++ app/pb_kits/playbook/pb_list/docs/index.js
@@ -1,2 +1,13 @@
 export { default as ListDefault } from './_list_default.jsx'
 export { default as ListDark } from './_list_dark.jsx'
+export { default as ListXpadding } from './_list_xpadding.jsx'
+export { default as ListDarkBorderless } from './_list_dark_borderless.jsx'
+export { default as ListDarkLg } from './_list_dark_lg.jsx'
+export { default as ListDarkXpadding } from './_list_dark_xpadding.jsx'
+export { default as ListDarkOrdered } from './_list_dark_ordered.jsx'
+export { default as ListLayoutRight } from './_list_layout_right.jsx'
+export { default as ListLayoutLeft } from './_list_layout_left.jsx'
+export { default as ListLg } from './_list_lg.jsx'
+export { default as ListOrdered } from './_list_ordered.jsx'
+export { default as ListBorderless } from './_list_borderless.jsx'
+